+++ app/src/main/kotlin/net/turtton/ytalarm/viewmodel/PlaylistViewModel.kt
@@ -18,15 +18,15 @@ class PlaylistViewModel(private val repository: DataRepository) : ViewModel() {
         repository.getAllPlaylistsSync()
     }
 
-    fun getFromId(id: Int): LiveData<Playlist> {
+    fun getFromId(id: Long): LiveData<Playlist> {
         return repository.getPlaylistFromId(id).asLiveData()
     }
 
-    fun getFromIdAsync(id: Int): Deferred<Playlist?> = viewModelScope.async {
+    fun getFromIdAsync(id: Long): Deferred<Playlist?> = viewModelScope.async {
         repository.getPlaylistFromIdSync(id)
     }
 
-    fun getFromIdsAsync(ids: List<Int>): Deferred<List<Playlist>> = viewModelScope.async {
+    fun getFromIdsAsync(ids: List<Long>): Deferred<List<Playlist>> = viewModelScope.async {
         repository.getPlaylistFromIdsSync(ids)
     }
